separate source, tests and script
